Fix type for stdhdev and pdop

+++ src/org/openstreetmap/josm/data/gpx/GpxImageEntry.java
@@ -894,13 +894,13 @@ public class GpxImageEntry implements Comparable<GpxImageEntry>, IQuadBucketType
             }
             Double hposerr = getExifHPosErr();
             if (hposerr != null) {
-                wpt.put(GpxConstants.PT_STD_HDEV, hposerr);
+                wpt.put(GpxConstants.PT_STD_HDEV, hposerr.floatValue());
             }
             Double exifGpsDop = getExifGpsDop();
             if (exifGpsDop != null) {
                 // TODO check if the image GpsMeasureMode is 2D or 3D 
                 // to put the value to PT_HDOP or PT_PDOP 
-                wpt.put(GpxConstants.PT_PDOP, exifGpsDop);
+                wpt.put(GpxConstants.PT_PDOP, exifGpsDop.floatValue());
             }
             //TODO add gpsDiffMode gpstrack exifGpsProcMethod here?
         }
